Backend: Bookings db and implementation

See merge request I522311/rentride_be!22

+++ src/main/java/nl/fontys/s3/rentride_be/domain/booking/Booking.java
@@ -25,6 +25,6 @@ public class Booking {
     private City endCity;
     private Car car;
     private User user;
-    private Long distance;
-    private Long totalPrice;
+    private double distance;
+    private double totalPrice;
 }
